move reprovide under core
things under exchange implement the exchange interface. do we want to
deviate from that norm?
